+++ protocol/go/entityresolution/entity_resolution.pb.gw.go
@@ -57,6 +57,32 @@ func local_request_EntityResolutionService_ResolveEntities_0(ctx context.Context
 
 }
 
+func request_EntityResolutionService_CreateEntityChainFromJwt_0(ctx context.Context, marshaler runtime.Marshaler, client EntityResolutionServiceClient, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
+	var protoReq CreateEntityChainFromJwtRequest
+	var metadata runtime.ServerMetadata
+
+	if err := marshaler.NewDecoder(req.Body).Decode(&protoReq); err != nil && err != io.EOF {
+		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
+	}
+
+	msg, err := client.CreateEntityChainFromJwt(ctx, &protoReq, grpc.Header(&metadata.HeaderMD), grpc.Trailer(&metadata.TrailerMD))
+	return msg, metadata, err
+
+}
+
+func local_request_EntityResolutionService_CreateEntityChainFromJwt_0(ctx context.Context, marshaler runtime.Marshaler, server EntityResolutionServiceServer, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
+	var protoReq CreateEntityChainFromJwtRequest
+	var metadata runtime.ServerMetadata
+
+	if err := marshaler.NewDecoder(req.Body).Decode(&protoReq); err != nil && err != io.EOF {
+		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
+	}
+
+	msg, err := server.CreateEntityChainFromJwt(ctx, &protoReq)
+	return msg, metadata, err
+
+}
+
 // RegisterEntityResolutionServiceHandlerServer registers the http handlers for service EntityResolutionService to "mux".
 // UnaryRPC     :call EntityResolutionServiceServer directly.
 // StreamingRPC :currently unsupported pending https://github.com/grpc/grpc-go/issues/906.
@@ -85,6 +111,31 @@ func RegisterEntityResolutionServiceHandlerServer(ctx context.Context, mux *runt
 		}
 
 		forward_EntityResolutionService_ResolveEntities_0(annotatedContext, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)
+
+	})
+
+	mux.Handle("POST", pattern_EntityResolutionService_CreateEntityChainFromJwt_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
+		ctx, cancel := context.WithCancel(req.Context())
+		defer cancel()
+		var stream runtime.ServerTransportStream
+		ctx = grpc.NewContextWithServerTransportStream(ctx, &stream)
+		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
+		var err error
+		var annotatedContext context.Context
+		annotatedContext, err = runtime.AnnotateIncomingContext(ctx, mux, req, "/entityresolution.EntityResolutionService/CreateEntityChainFromJwt", runtime.WithHTTPPathPattern("/entityresolution/entitychain"))
+		if err != nil {
+			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
+			return
+		}
+		resp, md, err := local_request_EntityResolutionService_CreateEntityChainFromJwt_0(annotatedContext, inboundMarshaler, server, req, pathParams)
+		md.HeaderMD, md.TrailerMD = metadata.Join(md.HeaderMD, stream.Header()), metadata.Join(md.TrailerMD, stream.Trailer())
+		annotatedContext = runtime.NewServerMetadataContext(annotatedContext, md)
+		if err != nil {
+			runtime.HTTPError(annotatedContext, mux, outboundMarshaler, w, req, err)
+			return
+		}
+
+		forward_EntityResolutionService_CreateEntityChainFromJwt_0(annotatedContext, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)
 
 	})
 
@@ -151,13 +202,39 @@ func RegisterEntityResolutionServiceHandlerClient(ctx context.Context, mux *runt
 
 	})
 
+	mux.Handle("POST", pattern_EntityResolutionService_CreateEntityChainFromJwt_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
+		ctx, cancel := context.WithCancel(req.Context())
+		defer cancel()
+		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
+		var err error
+		var annotatedContext context.Context
+		annotatedContext, err = runtime.AnnotateContext(ctx, mux, req, "/entityresolution.EntityResolutionService/CreateEntityChainFromJwt", runtime.WithHTTPPathPattern("/entityresolution/entitychain"))
+		if err != nil {
+			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
+			return
+		}
+		resp, md, err := request_EntityResolutionService_CreateEntityChainFromJwt_0(annotatedContext, inboundMarshaler, client, req, pathParams)
+		annotatedContext = runtime.NewServerMetadataContext(annotatedContext, md)
+		if err != nil {
+			runtime.HTTPError(annotatedContext, mux, outboundMarshaler, w, req, err)
+			return
+		}
+
+		forward_EntityResolutionService_CreateEntityChainFromJwt_0(annotatedContext, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)
+
+	})
+
 	return nil
 }
 
 var (
 	pattern_EntityResolutionService_ResolveEntities_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1}, []string{"entityresolution", "resolve"}, ""))
+
+	pattern_EntityResolutionService_CreateEntityChainFromJwt_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1}, []string{"entityresolution", "entitychain"}, ""))
 )
 
 var (
 	forward_EntityResolutionService_ResolveEntities_0 = runtime.ForwardResponseMessage
+
+	forward_EntityResolutionService_CreateEntityChainFromJwt_0 = runtime.ForwardResponseMessage
 )