+++ service/internal/logger/logger.go
@@ -9,6 +9,7 @@ import (
 
 type Logger struct {
 	*slog.Logger
+	Audit *AuditLogger
 }
 
 type Config struct {
@@ -17,10 +18,34 @@ type Config struct {
 	Type   string `yaml:"type" default:"json"`
 }
 
-func (l *Logger) With(key string, value string) *Logger {
-	return &Logger{
-		Logger: l.Logger.With(key, value),
+// From the Slog docs (https://betterstack.com/community/guides/logging/logging-in-go/#customizing-slog-levels):
+// The log/slog package provides four log levels by default, with each one
+// associated with an integer value: DEBUG (-4), INFO (0), WARN (4), and ERROR (8).
+const (
+	// Currently setting AUDIT level to 10, a level above ERROR so it is always logged
+	LevelAudit = slog.Level(10)
+)
+
+var CustomLevelNames = map[slog.Leveler]string{
+	LevelAudit: "AUDIT",
+}
+
+// Used to support custom log levels showing up with custom labels as well
+// see https://betterstack.com/community/guides/logging/logging-in-go/#creating-custom-log-levels
+func customReplaceAttributes(_ []string, a slog.Attr) slog.Attr {
+	if a.Key == slog.LevelKey {
+		level, ok := a.Value.Any().(slog.Level)
+		if !ok {
+			return a
+		}
+		levelLabel, exists := CustomLevelNames[level]
+		if !exists {
+			levelLabel = level.String()
+		}
+		a.Value = slog.StringValue(levelLabel)
 	}
+
+	return a
 }
 
 func NewLogger(config Config) (*Logger, error) {
@@ -39,20 +64,40 @@ func NewLogger(config Config) (*Logger, error) {
 	switch config.Type {
 	case "json":
 		j := slog.NewJSONHandler(w, &slog.HandlerOptions{
-			Level: level,
+			Level:       level,
+			ReplaceAttr: customReplaceAttributes,
 		})
 		logger = slog.New(j)
 	case "text":
 		t := slog.NewTextHandler(w, &slog.HandlerOptions{
-			Level: level,
+			Level:       level,
+			ReplaceAttr: customReplaceAttributes,
 		})
 		logger = slog.New(t)
 	default:
 		return nil, fmt.Errorf("invalid logger type: %s", config.Type)
 	}
+
+	// Audit logger will always log at the AUDIT level and be JSON formatted
+	auditLoggerHandler := slog.NewJSONHandler(w, &slog.HandlerOptions{
+		Level:       LevelAudit,
+		ReplaceAttr: customReplaceAttributes,
+	})
+
+	auditLoggerBase := slog.New(auditLoggerHandler)
+	auditLogger := CreateAuditLogger(*auditLoggerBase)
+
 	return &Logger{
 		Logger: logger,
+		Audit:  auditLogger,
 	}, nil
+}
+
+func (l *Logger) With(key string, value string) *Logger {
+	return &Logger{
+		Logger: l.Logger.With(key, value),
+		Audit:  l.Audit.With(key, value),
+	}
 }
 
 func getWriter(config Config) (io.Writer, error) {