+++ service/internal/security/crypto_provider.go
@@ -2,14 +2,6 @@ package security
 
 import "crypto"
 
-type Config struct {
-	Type string `yaml:"type" default:"standard"`
-	// HSMConfig is the configuration for the HSM
-	HSMConfig HSMConfig `yaml:"hsm,omitempty" mapstructure:"hsm"`
-	// StandardConfig is the configuration for the standard key provider
-	StandardConfig StandardConfig `yaml:"standard,omitempty" mapstructure:"standard"`
-}
-
 type CryptoProvider interface {
 	RSAPublicKey(keyID string) (string, error)
 	RSAPublicKeyAsJSON(keyID string) (string, error)
@@ -20,15 +12,4 @@ type CryptoProvider interface {
 	GenerateEphemeralKasKeys() (any, []byte, error)
 	GenerateNanoTDFSessionKey(privateKeyHandle any, ephemeralPublicKey []byte) ([]byte, error)
 	Close()
-}
-
-func NewCryptoProvider(cfg Config) (CryptoProvider, error) {
-	switch cfg.Type {
-	case "hsm":
-		return New(&cfg.HSMConfig)
-	case "standard":
-		return NewStandardCrypto(cfg.StandardConfig)
-	default:
-		return NewStandardCrypto(cfg.StandardConfig)
-	}
 }